requirements.txt
annotated-types==0.7.0
arrow==1.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
flexcache==0.3
flexparser==0.4
idna==3.10
Pint==0.24.4
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
stravalib==2.4
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
